+++ CHANGELOG.rst
@@ -1,8 +1,15 @@
 Changelog for Nautilus
 ======================
 
+0.1.38 (unreleased)
+-------------------
 
-0.1.37 (unreleased)
+Changed:
+^^^^^^^^
+- Set shader thresholds to display blue if pixel is < 1.0 and red if pixel is >= max threshold value
+
+
+0.1.37 (2023-09-12)
 -------------------
 
 Added:
@@ -12,6 +19,12 @@ Added:
 - Ability to specify the type of data being recorded. Current options are ``Calcium Imaging``
   and ``Voltage Imaging``
 - Ability to disable Live View while acquisition is running
+- Warn users when no platemap has been selected when starting acquisition
+
+Changed:
+^^^^^^^^
+- Turn off live view and LED light when acquisition stops
+- Update plate format files
 
 Fixed:
 ^^^^^^